requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
google-api-python-client
google-auth
google-auth-oauthlib